changed timing from seconds to milliseconds

+++ qrazor-fx3/src/xreal/g_func.h
@@ -52,7 +52,7 @@ struct moveinfo_t
 	float		decel;
 	float		distance;
 
-	float		wait;
+	int		wait;
 
 	// state data
 	int		state;
